Move private pages for users to /profile
We currently have /user/<username> as a namespace for user profiles. We
had a number of subpages of /user which contained tool for users (e.g.
import, reset timestamps, export). This would cause a problem if a user
registered a musicbrainz user with these names and tried to log into
listenbrainz. To fix this, move all user-related pages to /profile.
