+++ listenbrainz/tests/integration/test_user_views.py
@@ -1,12 +1,10 @@
-
-import sys
-import os
-import uuid
-from listenbrainz.tests.integration import IntegrationTestCase
-from flask import url_for
-import listenbrainz.db.user as db_user
-import time
 import json
+
+from flask import url_for
+
+import listenbrainz.db.user as db_user
+from listenbrainz.tests.integration import IntegrationTestCase
+
 
 class UserViewsTestCase(IntegrationTestCase):
 
@@ -24,23 +22,4 @@ class UserViewsTestCase(IntegrationTestCase):
             content_type = 'application/json'
         )
 
-    def test_export(self):
-        """
-        Test for the user export of ListenBrainz data.
-        """
-        # test get requests to export view first
-        self.temporary_login(self.user['id'])
-        resp = self.client.get(url_for('user.export_data'))
-        self.assert200(resp)
 
-        # send two listens for the user
-        resp = self.send_listens()
-        self.assert200(resp)
-
-        time.sleep(2)
-
-        # now export data and check if contains all the listens we just sent
-        resp = self.client.post(url_for('user.export_data'))
-        self.assert200(resp)
-        data = json.loads(resp.data)
-        self.assertEqual(len(data), 3)
