+++ listenbrainz/webserver/__init__.py
@@ -134,8 +134,10 @@ def _register_blueprints(app):
     from listenbrainz.webserver.views.api import api_bp
     from listenbrainz.webserver.views.api_compat import api_bp as api_bp_compat
     from listenbrainz.webserver.views.user import user_bp
+    from listenbrainz.webserver.views.profile import profile_bp
     app.register_blueprint(index_bp)
     app.register_blueprint(login_bp, url_prefix='/login')
     app.register_blueprint(user_bp, url_prefix='/user')
+    app.register_blueprint(profile_bp, url_prefix='/profile')
     app.register_blueprint(api_bp, url_prefix=API_PREFIX)
     app.register_blueprint(api_bp_compat)
