+++ listenbrainz/webserver/templates/index/lastfm-proxy.html
@@ -23,7 +23,7 @@
       Clients supporting the old version of the <a href="http://www.audioscrobbler.net/development/protocol/">
       AudioScrobbler API</a> (such as VLC and Spotify) can be configured to work with ListenBrainz by making the client point
       to <a href="http://proxy.listenbrainz.org">http://proxy.listenbrainz.org</a> and using MusicBrainz ID as username and the 
-      <a href="{{ url_for('user.info') }}">LB auth token</a> as password. 
+      <a href="{{ url_for('profile.info') }}">LB auth token</a> as password.
     </p>
 
     <p>
