+++ listenbrainz/webserver/templates/navbar.html
@@ -28,7 +28,7 @@
       {# Not showing user stuff on error pages. If attempt to load user info fails there, page will not render. #}
       {% if not error %}
         <ul class="nav navbar-nav navbar-right">
-          <li class="import-link"><a href="{{ url_for('index.import_data') }}">Import!</a></li>
+          <li class="import-link"><a href="{{ url_for('profile.import_data') }}">Import!</a></li>
           {% if not current_user or current_user.is_anonymous() %}
             <li><a href="{{ url_for('login.index') }}">Sign in</a></li>
           {% else %}
@@ -38,9 +38,9 @@
               </a>
               <ul class="dropdown-menu" role="menu">
                 <li><a href="{{ url_for('user.profile', user_name=current_user.musicbrainz_id) }}">Your Listens</a></li>
-                <li><a href="{{ url_for('user.import_data') }}">Import listens</a></li>
-                <li><a href="{{ url_for('user.export_data') }}">Export listens</a></li>
-                <li><a href="{{ url_for('user.info') }}">Your Info</a></li>
+                <li><a href="{{ url_for('profile.import_data') }}">Import listens</a></li>
+                <li><a href="{{ url_for('profile.export_data') }}">Export listens</a></li>
+                <li><a href="{{ url_for('profile.info') }}">Your Info</a></li>
                 <li><a href="{{ url_for('login.logout') }}">Sign out</a></li>
               </ul>
             </li>
