+++ listenbrainz/webserver/templates/profile/info.html
@@ -10,7 +10,7 @@
 
     <p>If you want to reset your previous import timestamp, click below</p>
     <p>
-      <span class="btn btn-info btn-lg" style="width: 300px"><a href="{{ url_for("user.reset_latest_import_timestamp") }}" style="color: white;">Reset Import Timestamp</a></span>
+      <span class="btn btn-info btn-lg" style="width: 300px"><a href="{{ url_for("profile.reset_latest_import_timestamp") }}" style="color: white;">Reset Import Timestamp</a></span>
     </p>
 
     <h3>User token</h3>
@@ -25,7 +25,7 @@
 
     <p>If you want to reset your token, click below</p>
     <p>
-      <span class="btn btn-info btn-lg" style="width: 200px"><a href="{{ url_for("user.reset_token") }}" style="color: white;">Reset token</a></span>
+      <span class="btn btn-info btn-lg" style="width: 200px"><a href="{{ url_for("profile.reset_token") }}" style="color: white;">Reset token</a></span>
     </p>
 
 {% endblock %}
