+++ listenbrainz/webserver/templates/user/export.html
@@ -11,7 +11,7 @@
         Export your listen history in JSON format and download it. Click Download to proceed.
     </p>
     <p>
-      <form action="{{url_for("user.export_data")}}" method='POST'>
+      <form action="{{url_for("profile.export_data")}}" method='POST'>
         <button class="btn btn-warning btn-lg" type="submit">Download</button>
       </form>
     </p>
