+++ listenbrainz/webserver/templates/user/import.html
@@ -44,12 +44,12 @@
     </p>
     <div class="well">
     {% if alpha_import_status == "WAITING" %}
-      <p> <strong> Your request to import data from ListenBrainz alpha is in a queue and we will get to it as soon as possible! You can reload this page or click <a href = {{ url_for('user.import_data') }}>here</a> to know the status of the request. </strong> </p>
+      <p> <strong> Your request to import data from ListenBrainz alpha is in a queue and we will get to it as soon as possible! You can reload this page or click <a href = {{ url_for('profile.import_data') }}>here</a> to know the status of the request. </strong> </p>
     {% elif alpha_import_status == "DONE" %}
       <p><strong>Your listen import from ListenBrainz Alpha has completed successfully!</strong></p>
     {% endif %}
     {% if alpha_import_status == "NO_REQUEST" or alpha_import_status == "DONE" %}
-      <a href={{ url_for("user.import_from_alpha") }} style="color: white">
+      <a href={{ url_for("profile.import_from_alpha") }} style="color: white">
         <p>
           <span class="btn btn-warning btn-lg" style="width: 400px">Import Now!</span>
         </p>
