+++ listenbrainz/webserver/views/test/test_user.py
@@ -1,8 +1,9 @@
-from listenbrainz.webserver.testing import ServerTestCase
 from flask import url_for
+
 import listenbrainz.db.user as db_user
 from listenbrainz.db.testing import DatabaseTestCase
-import time
+from listenbrainz.webserver.testing import ServerTestCase
+
 
 class UserViewsTestCase(ServerTestCase, DatabaseTestCase):
     def setUp(self):
@@ -10,6 +11,10 @@ class UserViewsTestCase(ServerTestCase, DatabaseTestCase):
         DatabaseTestCase.setUp(self)
         self.user = db_user.get_or_create('iliekcomputers')
         self.weirduser = db_user.get_or_create('weird\\user name')
+
+    def tearDown(self):
+        ServerTestCase.tearDown(self)
+        DatabaseTestCase.tearDown(self)
 
     def test_user_page(self):
         response = self.client.get(url_for('user.profile', user_name=self.user['musicbrainz_id']))
@@ -36,44 +41,3 @@ class UserViewsTestCase(ServerTestCase, DatabaseTestCase):
         )
         self.assert200(response)
         self.assertIn('var user_name = "weird%5Cuser%20name";', response.data.decode('utf-8'))
-
-    def test_reset_import_timestamp_get(self):
-        self.temporary_login(self.user['id'])
-        response = self.client.get(url_for('user.reset_latest_import_timestamp'))
-        self.assertTemplateUsed('user/resetlatestimportts.html')
-        self.assert200(response)
-
-    def test_reset_import_timestamp_post(self):
-        self.temporary_login(self.user['id'])
-        val = int(time.time())
-        db_user.update_latest_import(self.user['musicbrainz_id'], val)
-
-        response = self.client.post(
-            url_for('user.reset_latest_import_timestamp'),
-            data={
-                'reset': 'yes',
-                'token': self.user['auth_token']
-            }
-        )
-        self.assertStatus(response, 302) # should have redirected to the import page
-        self.assertRedirects(response, url_for('user.import_data'))
-        ts = db_user.get(self.user['id'])['latest_import'].strftime('%s')
-        self.assertEqual(int(ts), 0)
-
-    def test_user_info_view(self):
-        """Tests the user info view and makes sure auth token is present there"""
-        self.temporary_login(self.user['id'])
-        response = self.client.get(url_for('user.info', user_name=self.user['musicbrainz_id']))
-        self.assertTemplateUsed('user/info.html')
-        self.assert200(response)
-        self.assertIn(self.user['auth_token'], response.data.decode('utf-8'))
-
-    def test_user_info_not_logged_in(self):
-        """Tests user info view when not logged in"""
-        response = self.client.get(url_for('user.info'))
-        self.assertStatus(response, 302)
-        self.assertRedirects(response, url_for('login.index', next='/user/info'))
-
-    def tearDown(self):
-        ServerTestCase.tearDown(self)
-        DatabaseTestCase.tearDown(self)
